+++ app/src/main/java/top/niunaijun/blackboxa/view/fake/FakeManagerActivity.kt
@@ -13,7 +13,6 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import cbfg.rvadapter.RVAdapter
-import com.afollestad.materialdialogs.MaterialDialog
 import top.niunaijun.blackbox.entity.location.BLocation
 import top.niunaijun.blackbox.fake.frameworks.BLocationManager
 import top.niunaijun.blackboxa.R
@@ -22,6 +21,7 @@ import top.niunaijun.blackboxa.databinding.ActivityListBinding
 import top.niunaijun.blackboxa.util.InjectionUtil
 import top.niunaijun.blackboxa.util.inflate
 import top.niunaijun.blackboxa.util.toast
+import top.niunaijun.blackboxa.view.base.DsDialogs
 import top.niunaijun.blackboxa.view.base.BaseActivity
 
 
@@ -64,16 +64,16 @@ class FakeManagerActivity : BaseActivity() {
     }
 
     private fun disableFakeLocation(item: FakeLocationBean,position:Int) {
-        MaterialDialog(this).show {
-            title(R.string.close_fake_location)
-            message(text = getString(R.string.close_app_fake_location,item.name))
-            negativeButton(R.string.cancel)
-            positiveButton(R.string.done){
+        DsDialogs.confirm(
+            context = this,
+            title = R.string.close_fake_location,
+            message = getString(R.string.close_app_fake_location, item.name),
+            positive = R.string.done
+        ) {
                 BLocationManager.disableFakeLocation(currentUserID(),item.packageName)
                 toast(getString(R.string.close_fake_location_success,item.name))
                 item.fakeLocationPattern = BLocationManager.CLOSE_MODE
                 mAdapter.replaceAt(position,item)
-            }
         }
     }